Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int d,a,b,c;
-    scanf("%d %d %d",&a,&b,&c);
+    float d,a,b,c;
+    scanf("%f %f %f",&a,&b,&c);
     d=((a+b+c)/2);
     printf("Average: %.2f",d);
     return 0;
